Update database schema to support same movie from multiple websites.
Add an extra table "data_sources" where to store all the sources of the same movie with id "movie_id". This way, indexing and retrieving documents have no duplicates, but each movie still has the link to all the websites where it was scraped from.
When a crawler tries to save a movie that exists already, the movie is updated with the most relevant data (i.e. coalescing and keeping the longest plot and description).