Improve indexing and add support for server-side cursor.
After reviews were added, the size of the documents retrieved for indexing grew.
For this reason, loading in memory all the information may not be possible and make the program crash; this can be solved by retrieving the information of batches (currently of 1000 movies), which can progressively build the index.